+++ Assets/LDtkUnity/Runtime/Fields/LDtkFields.cs
@@ -14,6 +14,7 @@ namespace LDtkUnity
     public partial class LDtkFields : MonoBehaviour
     {
         internal const string PROPERTY_FIELDS = nameof(_fields);
+        
         [SerializeField] private LDtkField[] _fields = Array.Empty<LDtkField>();
         
         private readonly Dictionary<string, int> _keys = new Dictionary<string, int>();
@@ -27,8 +28,7 @@ namespace LDtkUnity
         {
             for (int i = 0; i < _fields.Length; i++)
             {
-                LDtkField field = _fields[i];
-                _keys.Add(field.Identifier, i);
+                _keys.Add(_fields[i].Identifier, i);
             }
         }