+++ Assets/LDtkUnity/Runtime/Tools/Bank/LDtkIidComponentBank.cs
@@ -59,7 +59,7 @@ namespace LDtkUnity
         }
         
         /// <summary>
-        /// Gets a iid GameObject. The objects are only available after their OnEnable. Otherwise, try using <see cref="FindObjectOfIid"/>. <br/>
+        /// Gets an iid GameObject. The objects are only available after their OnEnable. Otherwise, try using <see cref="FindObjectOfIid"/>. <br/>
         /// If the component was not found, then make sure the referenced component is active and accessed after it's OnEnable.
         /// </summary>
         /// <param name="iid">